requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-socks==0.10.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
pillow==11.0.0
propcache==0.2.1
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-socks==2.5.3
requests==2.32.3
rsa==4.9
sniffio==1.3.1
Telethon==1.38.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.18.3
google-genai
